requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-WTF==1.1.1
mysqlclient==2.1.1
python-dotenv==1.0.0
Werkzeug==2.2.3
WTForms==3.0.1
email-validator==2.0.0
gunicorn==20.1.0 